chore: move modular to inside lib/ directory
