+++ packages/storage/lib/index.js
@@ -28,31 +28,6 @@ import { getGsUrlParts, getHttpUrlParts, handleStorageEvent } from './utils';
 import version from './version';
 import fallBackModule from './web/RNFBStorageModule';
 
-export {
-  getStorage,
-  connectStorageEmulator,
-  ref,
-  deleteObject,
-  getBlob,
-  getBytes,
-  getDownloadURL,
-  getMetadata,
-  getStream,
-  list,
-  listAll,
-  updateMetadata,
-  putFile,
-  writeToFile,
-  toString,
-  child,
-  setMaxDownloadRetryTime,
-  setMaxOperationRetryTime,
-  setMaxUploadRetryTime,
-  refFromURL,
-  uploadString,
-  uploadBytesResumable,
-} from '../modular/index';
-
 const namespace = 'storage';
 const nativeEvents = ['storage_event'];
 const nativeModuleName = 'RNFBStorageModule';
@@ -232,5 +207,7 @@ export default createModuleNamespace({
 // storage().X(...);
 // firebase.storage().X(...);
 export const firebase = getFirebaseRoot();
+
+export * from './modular/index';
 
 setReactNativeModule(nativeModuleName, fallBackModule);
